Create camera container with shared functionality for both cameras; fix tests
